Covariance Fixed
RR EnKF working pretty well. Found error with covariance in version before which was causing P to be too large. That is fixed. But currently trying to figure out why stat1 causes P to grow and stat2 causes P to decrease when they are both in the same sim... On there own (captured in this version with measurement sets that contain only stat1 and only stat2) they seem to be okay and stat1 is not causing P to grow, BUT stat2 has worse results than stat1 (percentages up to ~80%, whereas stat2 had always under 25%)
Next, I might try to propagate every 10 seconds even if there isn't a measurement.