ECI to ECEF Correction
Found an error in the eci to ecef conversion. I believe I fixed it everywhere. It did change soem inputs to filter_functions functions.
The error was that in some places it was just using simulation time to rotate earth (incorrect), and in other places it was using UT time (correct).